requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib==1.7.4
pyotp==2.9.0
qrcode==7.4.2
pillow==10.1.0
python-multipart==0.0.6
argon2-cffi==23.1.0
cryptography==41.0.5
pydantic-settings==2.0.3
